controller

+++ backend/controllers/product.controller.js
@@ -48,11 +48,11 @@ export const updateProduct = async (req, res) => {
     });
 
     // Check if the product exists
-    if (!updatedProduct) {
-      return res
-        .status(404)
-        .json({ success: false, message: "Product not found" });
-    }
+    // if (!updatedProduct) {
+    //   return res
+    //     .status(404)
+    //     .json({ success: false, message: "Product not found" });
+    // }
 
     res.status(200).json({ success: true, data: updatedProduct });
   } catch (error) {
@@ -63,11 +63,18 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  // Check if the ID is valid
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product id" });
+  }
+  
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product deleted" });
   } catch (error) {
     console.log("error in deleting products:", error.message);
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
